Display improvements
- Better live display of IFS
- Includes the new transform while it is being created
- Improved colour rendering and alpha compositing when viewing
- Uses HSB colour generation instead of LUT
- Tidied up detail view HTML for matrix coefficients using a table

There is still more work to be done on the colour view, in particular the compositing/alpha blending. I have added two system
properties, 'explorer.rule' and 'explorer.alpha' which can be set in $JAVA_OPTS to define the arguments used to create the
AlphaComposite used to draw individual points when rendering the IFS view. The defaults give a reasonable result, but it would
be interesting to revisit the actual drawing mechanism, perhaps setting individual pixel values in the BufferedImage rather than
filling in a tiny Rect.

TODO Update the documentation to incorporate these changes for 1.0.3 release